Upload the /static directory
Upload the /static directory, that stores the icon images locally. 

Notice that, in order to reduce the mini-program's package size, we delete all the icon images that are not referred by the scripts. 

In more details, the icon images in this directory are used by the `tabBar` of the `pages.json`